Move tasks to separate files
