+++ src/GramatykiGrafoweAGH/project/task1.py
