+++ src/GramatykiGrafoweAGH/project/task2.py
@@ -1,0 +1,13 @@
+import networkx as nx
+
+from GramatykiGrafoweAGH import production
+
+
+@production
+def P3(G: nx.Graph) -> None:
+    raise NotImplementedError()
+
+
+@production
+def P4(G: nx.Graph) -> None:
+    raise NotImplementedError()
