+++ src/GramatykiGrafoweAGH/project/task3.py
@@ -1,0 +1,13 @@
+import networkx as nx
+
+from GramatykiGrafoweAGH import production
+
+
+@production
+def P5(G: nx.Graph) -> None:
+    raise NotImplementedError()
+
+
+@production
+def P6(G: nx.Graph) -> None:
+    raise NotImplementedError()
