+++ src/GramatykiGrafoweAGH/project/task4.py
@@ -1,0 +1,8 @@
+import networkx as nx
+
+from GramatykiGrafoweAGH import production
+
+
+@production
+def P7(G: nx.Graph) -> None:
+    raise NotImplementedError()
