+++ src/GramatykiGrafoweAGH/project/task6.py
@@ -1,0 +1,8 @@
+import networkx as nx
+
+from GramatykiGrafoweAGH import production
+
+
+@production
+def P9(G: nx.Graph) -> None:
+    raise NotImplementedError()
